Update README.md
looking at the code it looks like the false value is intended to work this way, but doesn't seem to be documented.  I found out about it from http://stackoverflow.com/questions/12408015/prevent-items-to-be-droppable-in-other-parent-child-elements.  If this isn't intended behaviour, but just a side effect, then please dismiss this and indicate the "right" way to limit draggables to their existing container.