Enable preprocessing draggable from non-ko source
I am receiving draggables from other parts of the application which are not under my direct control. To make these interop with my ko-sortables, I need to add the data to the dom-element's `DRAGKEY` property before the `dragged` callback could be fired.
Since I did not want to change the existing behaviour of the `dragged` callback, I've added a new `accept` callback similar, but not equal, to the accept callback in droppable. if it returns false, the item will not be added. That way, I could add the data to `DRAGKEY` in the accept callback.